subtitle
1 fix decode some embedding subtitle crash

meetsdk
1 add frame_rate and birate in mediainfo
2 add getAudioSessionId()
3 support android arm64 build

Conflicts:
	android/MeetSDK/src/android/pplive/media/player/MeetVideoView.java
	engine2/build/android/jni/Android.mk
	engine2/player/audioplayer.cpp

+++ android/MeetSDK/src/android/pplive/media/subtitle/SimpleSubTitleParser.java
@@ -66,7 +66,7 @@ public class SimpleSubTitleParser implements Handler.Callback, SubTitleParser {
 	
 	private SubTitleSegment mSegment;
 	
-	private int mNativeContext;
+	private long mNativeContext;
 	private Handler mInnerHandler;
 	
 	public SimpleSubTitleParser() {
